refactor(runtime-api): extract auth helpers
